% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elx_consolidate.R
\name{elx_consolidate}
\alias{elx_consolidate}
\title{Consolidate SPARQL results for preset queries}
\usage{
elx_consolidate(sparql_response)
}
\description{
Consolidate SPARQL results for preset queries
}
