% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elx_council_votes.R
\name{elx_council_votes}
\alias{elx_council_votes}
\title{Retrieve Council votes on EU acts}
\usage{
elx_council_votes()
}
\value{
A data frame with Council votes on EU acts.
}
\description{
Executes a SPARQL query to the Council's endpoint.
}
\examples{
\donttest{
votes <- elx_council_votes()
}
}
