% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elx_parse_xml.R
\name{elx_parse_xml}
\alias{elx_parse_xml}
\title{Parse RDF/XML triplets to data frame}
\usage{
elx_parse_xml(sparql_response = "")
}
\arguments{
\item{sparql_response}{A valid response from the SPARQL endpoint}
}
\description{
An internal function to parse RDF/XML output from SPARQL queries.
}
