% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elx_run_query.R
\name{elx_run_query}
\alias{elx_run_query}
\title{Execute SPARQL queries}
\usage{
elx_run_query(resource_type, manual_type)
}
\arguments{
\item{resource_type}{Type of resource to be retrieved via SPARQL query, presets include "directive", "regulation" and "decision"}

\item{manual_type}{Define manually the type of resource to be retrieved}
}
\description{
Wrapper around SPARQL::SPARQL with pre-defined endpoint of EU Publications Office
Relies on elx_make_query to generate valid SPARQL queries
}
\examples{
elx_run_query(resource_type = "directive")
}
