% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPARQL.R
\docType{data}
\name{sparqlns}
\alias{sparqlns}
\title{SPARQL package functions with fixed namespaces to reduce dependency load}
\format{
An object of class \code{character} of length 1.
}
\usage{
sparqlns
}
\description{
SPARQL package functions with fixed namespaces to reduce dependency load
}
\keyword{datasets}
