% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elx_curia_list.R
\name{elx_curia_list}
\alias{elx_curia_list}
\title{Scrape list of court cases from Curia}
\usage{
elx_curia_list(data = c("all", "ecj_old", "ecj_new", "gc_all", "cst_all"))
}
\arguments{
\item{data}{Data to be scraped from four separate lists of cases maintained by Curia, defaults to "all"
which contains cases from Court of Justice, General Court and Civil Service Tribunal.}
}
\description{
Harvests data from lists of EU court cases from curia.europa.eu.
CELEX identifiers are extracted from hyperlinks where available.
}
\examples{
\donttest{
elx_curia_list(data = "cst_all")
}
}
