% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elx_make_query.R
\name{elx_make_query}
\alias{elx_make_query}
\title{Create SPARQL quries}
\usage{
elx_make_query(
  resource_type,
  manual_type = "",
  include_corrigenda = FALSE,
  order = TRUE
)
}
\arguments{
\item{resource_type}{Type of resource to be retrieved via SPARQL query, presets include "directive", "regulation" and "decision"}

\item{manual_type}{Define manually the type of resource to be retrieved}

\item{include_corrigenda}{Should the query include corrigenda? Defaults to FALSE}

\item{order}{Order results by ids, defaults to TRUE}
}
\description{
Generates pre-defined or manual SPARQL queries.
Mainly to be called internally by other functions.
List of available resource types: http://publications.europa.eu/resource/authority/resource-type
Note that not all resource types are compatible with the pre-defined query
}
\examples{
elx_make_query(resource_type = "directive")
elx_make_query(resource_type = "regulation")
elx_make_query(resource_type = "decision")
elx_make_query(resource_type = "manual", manual_type = "SWD")
}
