% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elx_make_query.R
\name{elx_make_query}
\alias{elx_make_query}
\title{Create SPARQL quries}
\usage{
elx_make_query(
  resource_type = c("directive", "regulation", "decision", "recommendation", "intagr",
    "caselaw", "manual", "proposal", "national_impl"),
  manual_type = "",
  include_corrigenda = FALSE,
  include_celex = TRUE,
  include_lbs = FALSE,
  include_date = FALSE,
  include_date_force = FALSE,
  include_date_endvalid = FALSE,
  include_date_transpos = FALSE,
  include_force = FALSE,
  include_eurovoc = FALSE,
  order = FALSE,
  limit = NULL
)
}
\arguments{
\item{resource_type}{Type of resource to be retrieved via SPARQL query}

\item{manual_type}{Define manually the type of resource to be retrieved}

\item{include_corrigenda}{If `TRUE`, results include corrigenda}

\item{include_celex}{If `TRUE`, results include CELEX identifier for each resource URI}

\item{include_lbs}{If `TRUE`, results include legal bases of legislation}

\item{include_date}{If `TRUE`, results include document date}

\item{include_date_force}{If `TRUE`, results include date of entry into force}

\item{include_date_endvalid}{If `TRUE`, results include date of end of validity}

\item{include_date_transpos}{If `TRUE`, results include date of transposition deadline for directives}

\item{include_force}{If `TRUE`, results include whether legislation is in force}

\item{include_eurovoc}{If `TRUE`, results include EuroVoc descriptors of subject matter}

\item{order}{Order results by ids}

\item{limit}{Limit the number of results, for testing purposes mainly}
}
\value{
A character string containing the SPARQL query
}
\description{
Generates pre-defined or manual SPARQL queries to retrieve document ids from Cellar.
List of available resource types: http://publications.europa.eu/resource/authority/resource-type .
Note that not all resource types are compatible with the pre-defined query.
}
\examples{
elx_make_query(resource_type = "directive", include_date = TRUE, include_force = TRUE)
elx_make_query(resource_type = "regulation", include_corrigenda = TRUE, order = TRUE)
elx_make_query(resource_type = "caselaw")
elx_make_query(resource_type = "manual", manual_type = "SWD")
}
